auth routes

+++ server/src/routes/memories.ts
@@ -1,10 +1,20 @@
 import { FastifyInstance } from 'fastify'
 import { z } from 'zod'
+
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+    const { sub: userId } = request.user
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -19,7 +29,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
+    const { sub: userId } = request.user
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -32,10 +44,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    if (!memory.isPublic && memory.userId !== userId) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
   app.post('/memories', async (request) => {
+    const { sub: userId } = request.user
+
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -49,14 +67,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '5ac8562f-5c67-49da-8319-5646cf8b8b6d',
+        userId,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
+    const { sub: userId } = request.user
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -71,7 +91,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== userId) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -85,12 +115,24 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
+    const { sub: userId } = request.user
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== userId) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
